mapleads/requirements.txt
curl_cffi
fastapi
uvicorn[standard]
aiosqlite
dnspython
python-dotenv
pytest
pytest-asyncio
httpx
